Update the plugin to use the new Python client (#35)
This updates the plugin to use the new Python client, which uses v2 of
the GraphQL endpoint and corresponding schema.

This introduces a new function, `read_annotation_file`, in the public
API of the package. This replaces `read_annotation_files` which was
mistakenly never exposed in the public API, so can be removed instead of
deprecated.

The new schema allows multiple tomograms per run (e.g. multiple
alignments, spacings/reconstructions), but I chose not to add extra
information to the name displayed because it's a little hard to
determine what those should be.

I tested the demo manually after these changes.

Closes https://github.com/chanzuckerberg/cryoet-data-portal/issues/1344